Add Digits for converting numbers to file strings
Rationale: when interrupting hold music with queue position,
if you use .PlayFile() | .Say() | .PlayFile()
then there is a pause in between each app execution in which
the hold-music leaks through.
This can be used to build up a file_string input to the playback
application which can be used for seamless playback over hold music.